syscalls/flock: open test file with O_RDWR mode
NFS requires the open mode to match the lock type, i.e. for LOCK_SH
file must be opened for read, and for LOCK_EX file must be opened
for write. Otherwise flock01 and flock04 fail on NFS with EBADF.

So open all test files in flock tests with O_RDWR mode that both
lock type requirements can be met.

Signed-off-by: Eryu Guan <eguan@redhat.com>
Acked-by: Jan Stancek <jstancek@redhat.com>