numa: fix memhog fails in test8
memhog test fails on first iteration (when the memhog binary is not
in the page cache) because the test script runs too fast and looks
at the statistics even befor the memmhog binary has a chance to run.

In this patch, just redirectiong the output to a file and then checking
the content in the script in a loop.

Suggested-by: Cyril Hrubis <chrubis@suse.cz>
Signed-off-by: Li Wang <liwang@redhat.com>
Tested-by: Cyril Hrubis <chrubis@suse.cz>