controllers/cpuset: do not test undefined behavior
The test_flags() function's first test for each file is an attempt
to write a space character into that file (NULL translates to " ").
The test expects that this attempt will be successful and that the flag
will get set to 0.

This behavior was changed in Linux kernel between versions 3.13.74 and
3.14. with the commit a742c59de66ea080afa3edaf3428b3cdd5aa87cd
"cgroup: unify cgroup_write_X64() and cgroup_write_string()".
With the new behavior, attempting to write a space character into
a flag file returns EINVAL; I find this behavior more consistent.

Flag files are an interface which is known to expect numeric values.
We already have a test in test_flags() which covers invalid input.
We should not attempt to write a space into a flag file because
the resulting behavior is not strictly defined anywhere.
Therefore, it would be best to drop the first test.

Signed-off-by: Martin Karac <martin.karac@oracle.com>
Acked-by: Jan Stancek <jstancek@redhat.com>