configure.ac: Generate linux_syscall_headers.h
This avoids a race in generating the linux_syscall_headers.h on a
parallel build with many CPUs.

Note that the race has been reduced in commit:

commit 3f385652efe811fe7491474f8513baf44cf0a12d
Author: Cyril Hrubis <chrubis@suse.cz>
Date:   Wed Jan 13 14:13:31 2016 +0100

    build: regen.sh (linux_syscall_numbers.h): Fix race

But was not completely fixed.

After discussing the problem the easiest solution is to generate the
header before the build processes starts otherwise we would have to make
complex changes to the LTP build system. So this commits piggy backs the
regen.sh script invocation at the end of the configure script.

Fixes #161

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Acked-by: Jan Stancek <jstancek@redhat.com>