mmap01: fix testcase on hppa architecture
The mmap() syscall to map a file at any given page-aligned address (via the
MAP_FIXED flag) can fail on the hppa architecture due to cache colouring
requirements of the architecture.

This patch drops the MAP_FIXED flag and thus maps the file instead at some
convenient address based on the hint as given by the addr parameter. The tests
covered by mmap01 check the *end* of the mmaped file for consistency, so
mapping the beginning of the file at any address is okay without breaking the
main intention of the mmap01 testcase.

Since we now map the file at some random address as provided by the kernel drop
the initialization of the to-be-mapped address as well.

Signed-off-by: Helge Deller <deller@gmx.de>
Acked-by: Jan Stancek <jstancek@redhat.com>