mmapstress01: Fix buffer overflow
The testcase was causing buffer overflow on temporary file creation if
the path to $PWD was long enough (limit was 64 characters).

Moreover the test was using absolute path to create files in $PWD, which
is not allowed in LTP, while it is even using tst_tmpdir()/tst_rmdir()
to create unique test temporary directory.

So the fix to the problem is to remove the code that constructs the path
and change file to be created in $PWD which points to the temporary
directory created by the test library.

Fixes #170.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>