lib/tst_test: Ignore unknown options for TCONF tests
The problem is that with TST_TEST_TCONF() the optional test options are
not propagated in the tst_test structure. As a matter of fact the test
library had no idea that the test was compiled out, which had to be
changed as well.

So this commit adds a string pointer to the tst_test structure, that if
set, causes the test library to ignore unknown options and exits the
test, as fast as possible, with a TCONF.

This fixes cases where a test was compiled out but the runtest file
passes some tests specific options which previously failed the test with
TBROK.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>