syscalls/setrlimit02: Remove the EFAULT test
Glibc has to convert between 32bit and 64bit struct rlimit in some cases (looks
like that happens when prlimit64 is not implemented in the kernel) which causes
this test to SegFault.

See sysdeps/unix/sysv/linux/setrlimit*.c for reference:

...

  if (rlimits->rlim_cur >= RLIM_INFINITY)
    rlimits32.rlim_cur = RLIM_INFINITY;
  else
    rlimits32.rlim_cur = rlimits->rlim_cur;
  if (rlimits->rlim_max >= RLIM_INFINITY)
    rlimits32.rlim_max = RLIM_INFINITY;
  else
    rlimits32.rlim_max = rlimits->rlim_max;

  return __setrlimit (resource, &rlimits32);

...

Hence we remove the EFAULT case from the test in order not to produce false
possitives.

Closes #193

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>