syscalls/mmap12: Do not fail on non-present pages
There are two problems with the testcase:

1. MAP_POPULATE is best effort operation.

   To quote Michal Hocko:

   "The semantic of MAP_POPULATE is rather vague and allows for nuances
   in future."

2. There is no guarantee that the pages will be present even if
   MAP_POPULATE caused read-ahead.

   To quote Michal again:

   "Ptes can be made !present, reclaimed or who knows what else in
   future yet that won't qualify as a regression. I find such a test
   questionable at best."

I still think that calling mmap() with MAP_POPULATE and checking that
the mapping is OK is a valid testcase itself, so instead of passing the
test on present pages we simply check that the mapping is zero-filled
(since we mapped empty file). I kept the page-present check still there
but it now produces only INFO messages, it could be removed though if
everyone agrees that it has no real value.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
CC: Michal Hocko <mhocko@suse.com>
Acked-by: Jan Stancek <jstancek@redhat.com>