device-drivers/zram01: Fix failures on ppc64le
The Btrfs filesystem on ppc64le needs at least 139MB due to larger page
size. Unfortunately the minimal size reported by mkfs.btrfs does not
seem to be correct in all cases so this patch simply bumps the device
size to 384MB which should be more than enough.

We also check that there is at least 1GB of RAM now and default to 25MB
and ext2 instead of Btrfs in order not to fail the test on a tiny
machines.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Acked-by: Jan Stancek <jstancek@redhat.com>