android: getcwd02: use temp dir from $TMPDIR if present
getcwd02 test are broken on an Android system as the device doesn't have
"/tmp", so make sure the test checks if a directory exists before using
it as a temporary directory. Look for an altternative path in 'TMPDIR'
environment variable and fallback to using that as the temporary
directory for the test if /tmp is not found.

Signed-off-by: Sandeep Patil <sspatil@google.com>
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>