syscalls/mbind01: Fix warnings without libnuma2
The #else branch should have to use TST_TEST_TCONF() which is the
standard way how to cope with compiled out testcases.

The hacky code that implements main() in a case that libnuma2 is not
preset produces "no retrun in non-void function" that is treated as an
error by modern build systems. Moreover it also breaks the option
parsing which will break the new runltp-ng testrunner in the future.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>