Smash size_t and off_t printf() formatting mistakes
This fixes bunch of warnings when LTP is compiled for 32bit where size_t
and ssize_t are 32bit but off_t is 64 bit.

Here is a simple guide how to print size_t and off_t:

* size_t -> %zu
* ssize_t -> %zi or %zd
* off_t -> %li + cast to (long)

Common mistake is to print the st_size from struct stat with %zu but
st_size is defined as off_t. Also don't forget that sizeof(foo) returns
size_t.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>