syscalls/[p]write: check for special case with NULL buffer and 0 nsize
As documented, calling write with 0 as nsize is valid and should
return 0.

Passing NULL as a buffer is also valid, and when combined with a size
of 0 has been shown historically to result in a returned value of 0.

SysV and BSD make a point to always return 0, regardless of the buffer
and therefore some software expects a call like write(fd, NULL, 0) to
never fail, so test for that combination as an exception to this test
that looks for an EFAULT.

The fact that this test was disabled for uclinux and a comment on
their implementation seems to suggest that this exception was known
before and was therefore clarified as suported later in the documentation.

Signed-off-by: Carlo Marcelo Arenas Belón <carlo@gmail.com>