syscalls/dup203: Fix make use of SAFE_CLOSE()
The patch:

commit df007542635aed8686b5ebc111b0807a67349b41
Author: Cyril Hrubis <chrubis@suse.cz>
Date:   Tue Oct 3 13:29:15 2017 +0200

    Make use of SAFE_CLOSE()

Broke the test since it opens and closes a fd in order to get free file
descriptor to pass to dup2() as newfd. The problem here is that
SAFE_CLOSE() sets the file descriptor to -1, which is usually good
thing.

This patch fixes the problem by using temporary variable to store the fd
to pass it to SAFE_CLOSE().

Fixes #219

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>