syscalls/madvise09: Fix TCONF propagation, remove .min_kver
This patch fixes TCONF propagation from the child to parent, the parent
have to handle the TCONF exit value.

And it also removes the .min_kver in order to enable the test on
distribution kernels that backported the MADV_FREE. Older kernels are
handled by the runtime EINVAL detection or will TCONF at the check for
memory cgroup support, so there is no real need to keep the .min_kver in
the source code.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Acked-by: Jan Stancek <jstancek@redhat.com>