containers: netns: load veth module
Most of regular distributions are shipping the virtual ethernet driver as
a kernel module. The test assumes it's built-in.
Attempt to load the veth module in the netns_setup() function.
It will allow to get the tests running on a distro kernel and avoid to
build a custom kernel for LTP purpose.

Signed-off-by: Fathi Boudra <fathi.boudra@linaro.org>