lib: Add tst_strstatus()
This function returns a string describing status as returned by 'wait()'
call. This is expected to be used when we want to TBROK a test if child
haven't terminated in an expected way.

The fucntion is not thread safe, since it uses static buffer, but I
doubt that we actually care in this case and making it thread-safe would
have complicated the nice and simple API.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>