ioctl: fix stty usage for busybox
The ioctl tests rely on stty to test if a tty is usable, and use the
long form "--file=" flag to tell stty which tty to test. Busybox doesn't
understand the long "--file=" flag, and this causes the LTP test harness
to skip tests:

> stty: invalid argument '--file=/dev/tty0'
> ioctl01_02    0  TINFO  :  Skipping ioctl01 with /dev/tty0

Luckily, both busybox stty and GNU stty understand the equivalent "-F"
flag. This patch makes LTP use this more common option.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Signed-off-by: Punit Agrawal <punit.agrawal@arm.com>