fzsync: Prevent early exit
In some tests it may be possible for the child thread to exit before it has
completed an equal number of iterations to main thread (the thread which sets
the exit state). Due to the way most tests use the library this is unlikely to
matter, but in theory it could cause some tests to exit uncleanly. To fix
this, we only exit if we find that the thread has progressed to a wait() call
which would spin indefinitely if exit were not set.

I already suspected this bug existed, but confirmed while trying to formally
verify the algorithm using Modex, see:
https://github.com/richiejp/modex_models

Signed-off-by: Richard Palethorpe <rpalethorpe@suse.com>