syscalls/keyctl06: update to test for follow-on fix
The original fix for the buffer overflow bug changed keyring_read() to
start returning a short value if the specified buffer is too small, but
the documented behavior is actually to return the size needed.  Update
the test to expect the documented behavior, as is provided by a
follow-on fix.

Signed-off-by: Eric Biggers <ebiggers@google.com>
Acked-by: Xiao Yang <yangx.jy@cn.fujitsu.com>
Tested-by: Li Wang <liwang@redhat.com>