direct_io: diotest4: use getpagesize() for mmap if shm header is absent
For some architectures, mmap(MAP_FIXED) has to be aligned to SHMLBA
(instead of getpagesize()) --

   https://sourceforge.net/p/ltp/mailman/ltp-list/?viewmonth=200804

This dependency on SHMLBA causes the test to fail building for Android
due to absence of sys/shm.h. Make the header dependency configurable and
use getpagesize() as mmap granularity if the header is absent allowing
the test to work on Android while not breaking the architectures that
depend on the shm headers.

Signed-off-by: Sandeep Patil <sspatil@google.com>
Acked-by: Jan Stancek <jstancek@redhat.com>