CVE-2012-0957: fix false report with UNAME26 personality
When uname() is called with UNAME26 personality, the
uname->release field will be filled twise by kernel. If the
length of uname->release initially filled by the normal value
of release is longer at least 2 bytes than the converted, the
test case will fail due to leaking the longer part, e.g,
4.9.65006+ (10-byte) is converted to 2.6.69+ (7-byte), and the
last 2 characters will be leaked and cause a failure of test.

In order to fix this issue, save the normal values returned by
the first call of uname(), and make a compare when checking
against the converted values returned by UNAME26 personality.

Signed-off-by: Jia Zhang <qianyue.zj@alibaba-inc.com>