thp: ensure hugetlbfs is available
The THP tests thp02 and thp03 depend on the presence of hugetlbfs to get
the default hugepage size. If hugetlbfs is not enabled, /proc/meminfo
does not contain "Hugepagesize" and the tests fail with -

safe_file_ops.c:220: BROK: Expected 1 conversions got 0 at thp03.c:85

which is hard to relate to the missing support for hugetlbfs.

Add tests for the presence of hugetlbfs for tests relying on this kernel
feature.

Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Punit Agrawal <punit.agrawal@arm.com>
Acked-by: Petr Vorel <pvorel@suse.cz>