sigwaitinfo01: catch SEGV and report success for bad_address2 testcase
The bad_address2 testcase passes (void *)1 as a sigset pointer to the
sigwaitinfo syscall. Unsurprisingly, this segfaults in libc rather than
returning -1 (errno = EFAULT) as LTP expects.

Instead, fork a child process which registers a default handler for
SIGSEGV to catch any SEGV's generated by the sigwaitinfo syscall. The
parent checks that the child was terminated by SIGSEGV to evaluate the
success of the test.

Signed-off-by: Punit Agrawal <punit.agrawal@arm.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>