migrate_pages02: update for 4.15
commit 313674661925 ("Unify migrate_pages and move_pages access checks")
tightened checks for unprivileged user. It requires that callers real
uid/gid matches target [es]uig/[es]gid and target to be dumpable
(see __ptrace_may_access()).

It broke the test, because change of only euid is no longer
sufficient.

This patch updates test to use setuid(), which sets all user IDs.
Since we can't change uid back to 0, we fork additional child
process for "caller", which is now free to drop privileges.

This change should be backwards-compatible with kernels < 4.15.
Tested on 4.15-rc6 and 4.14.

Signed-off-by: Jan Stancek <jstancek@redhat.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>