syscalls/inotify07: test inotify dir watch after drop caches
Add watch on an overlayfs lower directory, then copy up the directory
and drop dentry and inode caches.

An inotify watch pins the directory inode in cache, but not the dentry.
Execute operations on directory and child and expect events to be
reported on directory watch. This will fail if file system does
not obtain the pinned inode to the new allocated dentry after copy up
and drop caches.

This is a regression test for commit 31747eda41ef ("ovl: hash directory
inodes for fsnotify").

Signed-off-by: Amir Goldstein <amir73il@gmail.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>