syscalls/inotify08: test inotify file watch after drop caches
Add watch on an overlayfs lower file, then copy up the file
and drop dentry and inode caches.

An inotify watch pins the file inode in cache, but not the dentry.
Execute operations on file and expect events to be reported on file
watch. This will fail if file system does not obtain the pinned inode
to the new allocated dentry after copy up and drop caches.

This is a regression test for commit 764baba80168 ("ovl: hash non-dir by
lower inode for fsnotify").

Signed-off-by: Amir Goldstein <amir73il@gmail.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>