shell: ROD_SILENT: Print std{out,err} on failure
Recently we had a problem when openQA (test automation framework) passed
wrong device path as LTP_BIG_DEV. It has been much more complicated to
debug because the error message from mkfs was not shown at all, since
the tst_mkfs uses ROD_SILENT to execute the mkfs command.

The best solution for ROD_SILENT is to store the command output into a
variable and print it only on failure which is done by this patch.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Acked-by: Richard Palethorpe <rpalethorpe@suse.de>