tst_rod: Fix error propagation on 'command not found'
In a case that we try to execute nonexistent command we end up with
'command not found' in the test logs but the error is actually ignored.
That is because the pclose() returns the value as returned by wait() so
some of the upper bits are cleared on return from the main().

Since we are not actually using the return value for anything, we only
do tst_brk if it's non-zero we may as well normalize it with double
negation, which is what this patch does.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>