pids.sh: Fix spurious failures in test3 cleanup.
Sometimes the possitive testcases were failing in cleanup:

kill: (17773): No such process
 1 TBROK: kill -9 17773 failed

The reason is that when we fork pids_task1 the child is running in the
particual subgroup and hence it's pid is in the task file. And since the
child process did only exit(0) there was a race between the process exit
and the stop_pids_task function that read the tasks file.

The fix is obvious, we suspend the pids_task1 child to wait patiently
for it's execution.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>